backend/requirements.txt
fastapi
uvicorn[standard]
qdrant-client
httpx
python-dotenv
psycopg2-binary